Finish splitting display components
