+++ assets/src/components/apps/app/runbooks/runbook/display/DisplayBox.tsx
@@ -1,0 +1,18 @@
+import { Box } from 'grommet'
+
+import { recurse } from './misc'
+
+const border = ({ borderSize, borderSide, border }) => (
+  (borderSize || borderSide) ? { side: borderSide, color: border, size: borderSize } : border
+)
+
+export function DisplayBox({ children, attributes }) {
+  return (
+    <Box
+      {...(attributes || {})}
+      border={border(attributes)}
+    >
+      {recurse(children)}
+    </Box>
+  )
+}
