+++ assets/src/components/apps/app/runbooks/runbook/display/DisplayImage.tsx
@@ -1,0 +1,15 @@
+export function DisplayImage({
+  attributes: {
+    url, width = 250, height = 250, ...rest
+  },
+}) {
+  return (
+    <img
+      alt={url}
+      width={width}
+      height={height}
+      {...rest}
+      src={url}
+    />
+  )
+}
