+++ assets/src/components/apps/app/runbooks/runbook/display/DisplayInput.tsx
@@ -2,7 +2,9 @@ import { Input } from '@pluralsh/design-system'
 import { Div, Span } from 'honorable'
 import { useContext, useEffect, useState } from 'react'
 
-import { DisplayContext, convertType, valueFrom } from '../RunbookDisplay'
+import { DisplayContext } from '../RunbookDisplay'
+
+import { convertType, valueFrom } from './misc'
 
 export default function DisplayInput({ attributes, children }) {
   const { context, setContext, ...rest } = useContext(DisplayContext)
