+++ assets/src/components/apps/app/runbooks/runbook/display/DisplayMarkdown.tsx
@@ -1,0 +1,12 @@
+import { Markdown } from '@pluralsh/design-system'
+
+export function DisplayMarkdown({ attributes: { value, ...rest } }) {
+  if (!value) return null
+
+  return (
+    <Markdown
+      text={value}
+      {...rest}
+    />
+  )
+}
